% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarm.R
\name{jarm_fingerprint}
\alias{jarm_fingerprint}
\alias{print.jarm_result}
\title{Fingerprint a TLS server}
\usage{
jarm_fingerprint(host, port = 443L)

\method{print}{jarm_result}(x, ...)
}
\arguments{
\item{host}{host or IP to fingerprint}

\item{port}{port to connect on (defaults to \code{443L})}

\item{x}{\code{jarm_result} object}

\item{...}{unused}
}
\value{
\code{list} with class \code{jarm_result}

x (invisibly)
}
\description{
Fingerprint a TLS server
}
\examples{
jarm_fingerprint("rud.is")
}
\keyword{internal}
