% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarmed-package.R
\docType{package}
\name{jarmed}
\alias{jarmed}
\alias{jarmed-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/jarmed}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/jarmed/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
