% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarmed-package.R
\docType{package}
\name{jarmed}
\alias{jarmed}
\alias{jarmed-package}
\title{Fingerprint TLS Servers with Salesforece JARM Algorithm}
\description{
The Salesforce JARM Tool \url{https://github.com/salesforce/jarm} is an
active Transport Layer Security (TLS) server fingerprinting tool. JARM
fingerprints can be used to quickly verify that all servers in a group have
the same TLS configuration; group disparate servers on the internet by configuration,
identifying that a server may belong to Google vs. Salesforce vs. Apple, for example;
Identify default applications or infrastructure; and/or Identify malware command and
control infrastructure and other malicious servers on the Internet. Tools
are provided to generate JARM fingerprints.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/jarmed}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/jarmed/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
